pyproject.toml
[project]
name = "general-task-agent"
version = "1.0.0"
description = "A versatile AI agent capable of handling various types of tasks using the A2A protocol"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "uvicorn>=0.34.2",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "opentelemetry-api>=1.33.0",
    "opentelemetry-sdk>=1.33.0",
    "opentelemetry-exporter-otlp>=1.33.0",
    "opentelemetry-instrumentation-starlette>=0.54.0",
    "a2a-sdk>=0.2.6",
    "openai~=1.66.3",
    "tenacity~=9.0.0",
    "pyyaml~=6.0.2",
    "loguru~=0.7.3",
    "numpy",
    "datasets~=3.4.1",
    "fastapi~=0.115.11",
    "tiktoken~=0.9.0",
    "html2text~=2024.2.26",
    "gymnasium~=1.1.1",
    "pillow~=11.1.0",
    "browsergym~=0.13.3",
    "unidiff~=0.7.5",
    "browser-use~=0.1.40",
    "googlesearch-python~=1.3.0",
    "baidusearch~=1.0.3",
    "duckduckgo_search~=7.5.3",
    "aiofiles~=24.1.0",
    "pydantic_core~=2.33.1",
    "colorama~=0.4.6",
    "playwright~=1.51.0",
    "docker~=7.1.0",
    "pytest~=8.3.5",
    "pytest-asyncio~=0.25.3",
    "mcp~=1.5.0",
    "httpx>=0.27.0",
    "tomli>=2.0.0",
    "boto3~=1.37.18",
    "requests~=2.32.3",
    "beautifulsoup4~=4.13.3",
    "huggingface-hub~=0.29.2",
    "setuptools~=75.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 
